+++ celt/hybrid_encode_helpers_test.go
@@ -34,7 +34,7 @@ func TestApplyHybridPrefilterMatchesDisabledRunPrefilter(t *testing.T) {
 
 	prevPeriod := want.prefilterPeriod
 	prevGain := want.prefilterGain
-	pfResult := want.runPrefilter(wantPreemph, frameSize, want.TapsetDecision(), false, 0.91, 14, 0.22, 0.18, want.analysisMaxPitchRatio)
+	pfResult := want.runPrefilter(wantPreemph, frameSize, want.TapsetDecision(), false, 0.91, 14, 0.22, 0.18, float64(want.analysisMaxPitchRatio))
 	roundFloat64ToFloat32(wantPreemph)
 	want.lastPitchChange = false
 	if prevPeriod > 0 && (pfResult.gain > 0.4 || prevGain > 0.4) {
